fix/241025/fix-shortcomings
