+++ requirements.txt
@@ -4,3 +4,4 @@ pyqt5
 torch
 torchvision
 tqdm
+TensorFlow
